app/frontend/requirements.txt
streamlit
streamlit-option-menu
requests
pandas
python-dotenv
groq
openai
chromadb
psycopg2-binary
sqlalchemy
duckdb
plotly